requirements.txt
llama_index
google-generativeai
llama_index_llms_gemini
pypdf
python-dotenv
IPython
llama_index_embeddings_gemini
streamlit